requirements.txt
llama-index
llama-index-llms-openai
llama-index-embeddings-fastembed
llama-index-vector-stores-chroma
chromadb
pypdf
python-docx
python-pptx
streamlit
python-dotenv
pydantic-settings
openai
fastembed
streamlit-mic-recorder
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.100.0
firebase-admin==6.2.0
google-cloud-aiplatform==1.38.0